requirements.txt
asgiref==3.4.1
autopep8==1.5.7
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.6
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-heroku==0.3.1
filelock==3.6.0
gcloud==0.17.0
google-api-core==2.11.1
google-api-python-client==2.26.1
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.59.1
gunicorn==20.1.0
httplib2==0.20.1
idna==3.3
jws==0.1.3
oauth2client==3.0.0
oauthlib==3.1.1
Pillow==8.3.1
pipenv==2020.11.15
platformdirs==2.5.1
protobuf==4.23.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycryptodome==3.4.3
pyparsing==2.4.7
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2021.1
requests==2.11.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.13.1
virtualenv-clone==0.5.7
whitenoise==6.0.0
django-environ==0.10.0